+++ src/colour/palette_test.go
@@ -46,8 +46,8 @@ func TestPalette_GetRGB(t *testing.T) {
 	expected := [][]uint16{{0, 0, 0}, {65535, 65535, 65535}, {38731, 38731, 38731}, {0, 0, 0}}
 
 	for i, e := range expected {
-		if r, g, b := palette.GetRGB(byte(i)); r != e[0] || g != e[1] || b != e[2] {
-			t.Errorf("entry at %d not returned correctly: was [%d %d %d], expected %v", i, r, g, b, e)
+		if r, g, b := palette.GetRGB(byte(i)); uint16(r) != e[0] || uint16(g) != e[1] || uint16(b) != e[2] {
+			t.Errorf("entry at %d not returned correctly: was [%f %f %f], expected %v", i, r, g, b, e)
 		}
 	}
 }
@@ -68,7 +68,7 @@ func TestPalette_GetLitRGB(t *testing.T) {
 
 	for _, testCase := range testCases {
 		r, g, b := palette.GetLitRGB(testCase.index, testCase.lighting)
-		result := []uint16{r, g, b}
+		result := []uint16{uint16(r), uint16(g), uint16(b)}
 		if result[0] != testCase.expected[0] || result[1] != testCase.expected[1] || result[2] != testCase.expected[2] {
 			t.Errorf("index %d with lighting %f returned %v, expected %v", testCase.index, testCase.lighting, result, testCase.expected)
 		}
